Change ViewPager fragment lifecycle handling
Change the ViewPager so that it no longer holds all sections in memory. This has the effect that the fragment lifecycle can no longer be used as the application's lifecycle, so decouple the BaseElement lifecycles from the TabSectionFragment's lifecycles and manage them manually on our own. The onStop() cycles from ActivityListener is removed due to it not being useable anymore.