Add option to disable stability check
The stability check is very cautious in that it runs after EVERY boot,
not just the first one after a change is made. This means something as
simple as manually rebooting within two minutes after boot will ALWAYS
cause Synapse settings to be reverted, even if they had been running
stable for months. This change allows the user to manually prevent this
from happening once if they feel comfortable with their settings.